+++ src/BulletSoftBody/btSoftMultiBodyDynamicsWorld.cpp
@@ -121,7 +121,7 @@ void	btSoftMultiBodyDynamicsWorld::solveSoftBodiesConstraints( btScalar timeStep
 	}
 
 	// Solve constraints solver-wise
-	m_softBodySolver->solveConstraints((float) (timeStep * m_softBodySolver->getTimeScale() ));
+	m_softBodySolver->solveConstraints( timeStep * m_softBodySolver->getTimeScale() );
 
 }
